requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
huggingface_hub>=0.24.0
langchain-qdrant>=0.2.0
langgraph>=0.2.0
qdrant-client>=1.11.0
python-dotenv>=1.0.0
pypdf>=4.0.0
python-docx>=1.1.0
tiktoken>=0.7.0
rich>=13.0.0
langfuse>=2.0.0
